adding unit test

+++ EulerProject/Algorithms/1.cs
@@ -10,9 +10,9 @@ namespace EulerProject.Algorithms
     {
 
         //Find the sum of all the multiples of 3 or 5 below the provided parameter value number.
-        public static long multiplesOf3and5(int num)
+        public static int multiplesOf3and5(int num)
         {
-            long sum = 0;
+            int sum = 0;
             for (int i = 0; i < num; i++)
             {
                 if (i % 3 == 0 || i % 5 == 0) sum += i;
